+++ Sources/NukeExtensions/ImageViewExtensions.swift
@@ -3,7 +3,9 @@
 // Copyright (c) 2015-2026 Alexander Grebenyuk (github.com/kean).
 
 import Foundation
-import Nuke
+
+// NukeExtensions' API is written in Nuke types, so `import NukeExtensions` is enough to use it.
+@_exported import Nuke
 
 #if !os(macOS)
 import UIKit.UIImage